fix: move optional dependency to separate entry point
