+++ account-kit/infra/package.json
@@ -6,10 +6,10 @@
   "license": "MIT",
   "private": false,
   "type": "module",
-  "main": "./dist/esm/index.js",
-  "module": "./dist/esm/index.js",
-  "types": "./dist/types/index.d.ts",
-  "typings": "./dist/types/index.d.ts",
+  "main": "./dist/esm/exports/index.js",
+  "module": "./dist/esm/exports/index.js",
+  "types": "./dist/types/exports/index.d.ts",
+  "typings": "./dist/types/exports/index.d.ts",
   "sideEffects": false,
   "files": [
     "dist",
@@ -23,9 +23,14 @@
   ],
   "exports": {
     ".": {
-      "types": "./dist/types/index.d.ts",
-      "import": "./dist/esm/index.js",
-      "default": "./dist/esm/index.js"
+      "types": "./dist/types/exports/index.d.ts",
+      "import": "./dist/esm/exports/index.js",
+      "default": "./dist/esm/exports/index.js"
+    },
+    "./enhanced-apis": {
+      "types": "./dist/types/exports/enhancedApi.d.ts",
+      "import": "./dist/esm/exports/enhancedApi.js",
+      "default": "./dist/esm/exports/enhancedApi.js"
     },
     "./package.json": "./package.json"
   },
@@ -34,7 +39,7 @@
     "build": "yarn clean && yarn build:esm && yarn build:types",
     "build:esm": "tsc --project tsconfig.build.json --outDir ./dist/esm",
     "build:types": "tsc --project tsconfig.build.json --declarationDir ./dist/types --emitDeclarationOnly --declaration --declarationMap",
-    "fern:gen": "node ../../doc-gen/dist/esm/cli.js generate --in ./src/index.ts --out ../../docs/pages/reference/account-kit/infra",
+    "fern:gen": "node ../../doc-gen/dist/esm/cli.js generate --in ./src/exports/index.ts --out ../../docs/pages/reference/account-kit/infra",
     "clean": "rm -rf ./dist",
     "test": "vitest",
     "test:run": "vitest run"
