+++ account-kit/infra/src/enhancedApi/schema.ts
@@ -1,0 +1,4 @@
+import { Alchemy } from "alchemy-sdk";
+import z from "zod";
+
+export const AlchemySdkClientSchema = z.instanceof(Alchemy);
