+++ account-kit/infra/src/exports/enhancedApi.ts
@@ -1,0 +1,3 @@
+export type * from "../enhancedApi/schema.js";
+export type * from "../enhancedApi/alchemyEnhancedApisDecorator.js";
+export { alchemyEnhancedApiActions } from "../enhancedApi/alchemyEnhancedApisDecorator.js";
