--- account-kit/infra/src/index.ts
@@ -1,76 +1,0 @@
-export type * from "./actions/simulateUserOperationChanges.js";
-export { simulateUserOperationChanges } from "./actions/simulateUserOperationChanges.js";
-export type * from "./actions/types.js";
-export type * from "./alchemyTransport.js";
-export { alchemy, isAlchemyTransport } from "./alchemyTransport.js";
-export type * from "./chains.js";
-export {
-  arbitrum,
-  arbitrumGoerli,
-  arbitrumNova,
-  arbitrumSepolia,
-  base,
-  baseGoerli,
-  baseSepolia,
-  defineAlchemyChain,
-  fraxtal,
-  fraxtalSepolia,
-  goerli,
-  mainnet,
-  optimism,
-  optimismGoerli,
-  optimismSepolia,
-  polygon,
-  polygonAmoy,
-  polygonMumbai,
-  sepolia,
-  shape,
-  shapeSepolia,
-  worldChain,
-  worldChainSepolia,
-  zora,
-  zoraSepolia,
-  beraChainBartio,
-  opbnbMainnet,
-  opbnbTestnet,
-  soneiumMinato,
-  soneiumMainnet,
-  unichainMainnet,
-  unichainSepolia,
-  inkMainnet,
-  inkSepolia,
-  mekong,
-  monadTestnet,
-  openlootSepolia,
-  gensynTestnet,
-  riseTestnet,
-  storyMainnet,
-  storyAeneid,
-  celoAlfajores,
-  celoMainnet,
-  teaSepolia,
-  bobaSepolia,
-  bobaMainnet,
-} from "./chains.js";
-export type * from "./client/decorators/alchemyEnhancedApis.js";
-export { alchemyEnhancedApiActions } from "./client/decorators/alchemyEnhancedApis.js";
-export type * from "./client/decorators/smartAccount.js";
-export { alchemyActions } from "./client/decorators/smartAccount.js";
-export { isAlchemySmartAccountClient } from "./client/isAlchemySmartAccountClient.js";
-export type * from "./client/rpcClient.js";
-export { createAlchemyPublicRpcClient } from "./client/rpcClient.js";
-export type * from "./client/smartAccountClient.js";
-export { createAlchemySmartAccountClient } from "./client/smartAccountClient.js";
-export type * from "./client/types.js";
-export { getDefaultUserOperationFeeOptions } from "./defaults.js";
-export { getAlchemyPaymasterAddress } from "./gas-manager.js";
-export { alchemyFeeEstimator } from "./middleware/feeEstimator.js";
-export type * from "./middleware/gasManager.js";
-export * from "./alchemyTrackerHeaders.js";
-export {
-  alchemyGasManagerMiddleware,
-  alchemyGasAndPaymasterAndDataMiddleware,
-  type PolicyToken,
-} from "./middleware/gasManager.js";
-export { alchemyUserOperationSimulator } from "./middleware/userOperationSimulator.js";
-export type * from "./schema.js";
