+++ account-kit/infra/src/schema.ts
@@ -1,5 +1,4 @@
 import { ChainSchema } from "@aa-sdk/core";
-import { Alchemy } from "alchemy-sdk";
 import type { Chain } from "viem";
 import z from "zod";
 
@@ -8,5 +7,3 @@ export const AlchemyChainSchema = z.custom<Chain>((chain) => {
 
   return chain_.rpcUrls.alchemy != null;
 }, "chain must include an alchemy rpc url. See `defineAlchemyChain` or import a chain from `@account-kit/infra`.");
-
-export const AlchemySdkClientSchema = z.instanceof(Alchemy);
